reorganize the moroz service (#8)
Did a lot of cleaning up and reorganized code to follow a simple 1 file per endpoint model.
Each method of the santa.Service now corresponds to a svc_*.go file in the moroz package.
The implementation of the service method lives at the top, followed by boilerplate to create a go-kit endpoint and the
necessary http transport decode function. Finally any service middleware is written below.

The middleware.go file is for declaring middleware on santa.Service.

The server.go file is for registering HTTP routes with gorilla/mux.